Added keyboard support.
The state is contained in the Game type which currently only
contains a player, which only has 2D position. The player is
controlled via the arrow keys, fullscreen is toggled by alt+enter
and the game is exited with escape. The game can also still be
exited by closing the window.