Split different types into different modules.
After the last refactor it is now easier to move types around to
different modules. I also changed the representation of some types
to be more idiomatic of Nim. Some operators and iterators were
added to make it clearer how they are being used.

The most significant changes are how Maps are done. It was changed
from
```nim
\#old
Map = seq[seq[Tile]]

\#new
Map = object
  shape: tuple[w, h: Natural]
  data: seq[Tile]
```

This and some better access practices are provided, this is still
a work in progress so I don't need to expose as much of the
internals of the types.